Improvement: Introduce `TeamcityMessages::testFailed` overload
Before this patch, the method accepts `int` as duration having a
default value `-1` to indicate that tests duration is unknown.
Boost Testing framework (a callback) has duration parameter as
`unsigned long`. That was a cause of the warning (or an error in
the case of `-Werror`) with `-Wconversion`.

The patch offers the overload of the mentioned mehod without
`duration` parameter and turns `int` to `unsigned long` in the
other signature.